+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Markdown syntax highlighting in the item body editor (dark mode only, colors matched to Nova editor)
+
 ## [1.0.1] - 2026-02-20
 
 ### Fixed